improve ORM implementation
