+++ backend/koala/database/model/__init__.py
@@ -1,2 +1,6 @@
+from .entity import Entity, DataBaseEntity
+from .system import SystemID, System
+from .tool import ToolID, Tool
 from .document import DocumentID, Document
 from .link_docs_to_entity import LinkDocEntityID, LinkDocEntity
+from .link_system_to_tool import LinkSystemToolID, LinkSystemTool
